ImageGadget : Refactor with new private TileShader class
This is just shuffling code around really, but it does tie the LUT texture and associated GLSL shader together in a logical way, and stops us keeping textures alive after we've thrown away the shader they were intended for (when the next shader doesn't need a LUT). This alone probably isn't sufficient to justify the work, but we also need a place to store some additional data in an ABI compatible way for some forthcoming commits.