ShadingEngine : Don't read past the end of the data
The `ShadingEngine::shade()` interface isn't currently fit for purpose. It assumes that all VectorTypedData values are per-point, and everything else is constant. This falls down when OSLObject passes a constant PrimitiveVariable which happens to contain VectorTypedData, such as USD's `displayColor`. Without a change to the API we can't do better than avoiding reads past the end, which in the `displayColor` case is happily enough to do what people want. This is not sufficient for constant arrays with length greater than one, in which case we would need to allow the OSL shader to specify which index to read from the array, or to read the entire array in one go. That can wait though.