GafferUI : Fix Qt crashes when deleting windows during close events
This was fine in Qt 5.6 but crashes reliably in Qt 5.12. This seems to be due to the appearance of the QWindow and QWidgetWindow classes, which are associated with top-level QWidgets in Qt 5.12.

The sequence of events in `testRemoveOnCloseCrash()` is as follows :

1. CloseEvent is sent to the QWindow for our child `GafferUI.Window`.
2. `QWidgetWindow::event()` handles this by forwarding to the QWidget via `QWidgetPrivate::close_helper()`.
3. `close_helper()` sends the event to the top level QWidget, and it is handled in our `_WindowEventFilter`.
4. `_WindowEventFilter` forwards to `GafferUI.Window.close()`, which succeeds and emits `closedSignal()`.
5. The `removeOnClose` behaviour removes the child window from the parent, dropping the last reference to it in the process.
6. The child `GafferUI.Window` is destroyed, also destroying its associated QWidget. The QWidget destructor triggers the destruction of it's associated QWindow via `deleteTLSysExtra`. So far so good.
7. We return to `close_helper()` in 3. This appears to have been written carefully so that it can cope with the QWidget being deleted in close : it uses a QPointer to detect the deletion and avoid subsequent accesses.
8. We return to `QWidgetWindow::event()` in 2. This calls its base class `QWindow::event()` method, but only after calling `event.setAccepted( true )`.
9. `QWindow::event()` decides that since the close event was accepted, it really ought to call `destroy()` on itself. But that was done already in 6. Double destruction. Boom.

What's frustrating is that this pattern is clearly meant to be supported by `close_helper()`, and it worked fine in Qt 5.6. It's just gone south in Qt 5.12. The workaround is to use `keepUntilIdle()` to defer the destruction of the window.

I haven't included test cases for the CompoundEditor/NotificationMessageHandler fixes because they are much harder to test. I was able to reproduce reliable CompoundEditor crashes interatively though, and can confirm that they are now fixed. NotificationMessageHandler never reproduced a crash, but it fits the same pattern as everything else here so I'm including the fix out of an abundance of caution.