GafferUI : Use `keepUntilIdle()` where appropriate
This expresses what we're trying to achieve more clearly, and means that we can have the visible effects we want immediately (we are now only deferring resource destruction). There are other spots that may benefit from this treatment, but they are less trivial to implement and test, so can wait for such a time as we are refactoring that code anyway.