OpenColorIOTransform : Add `processor()` and `processorHash()` public methods
We had intended the `transform()` method to be useable for replicating the effects of the node in other settings, but that is inadequate as it doesn't take account of the OCIO context as specified by `contextPlug()`. It also didn't perform context substitutions for string plugs due to not being a `Gaffer::Process`.