TransformTool : Use standardised EditScope logic for Selection
We were inconsistent with the standardised logic established whilst working
on the `SceneViewInspector`. We picked EditScopes in preference over any
applicable nodes inside the scope. This could result in unnecessary
`TranformEdits` being made, even if the the target location's source node
was inside the scope, or there was an existing Transform node editing
that location.