ImageGadget : Fix display transform regressions
In 0.57, display transform plugs could be edited at any time, and this would be reflected in the ImageGadget. Display transforms could also have context-sensitive values and those would be respected too. This was lost when we baked the transform to a shader in `ImageGadget::setDisplayTransform()`. We now track display transform changes and update the shader appropriately, using a hash to avoid unnecessary updates.