SpreadsheetUI : Create common plug model and map via proxies
Previously, we used multiple copies of an adaptable model that provided
specific data for the Cell, RowNames and Defaults table views. In order
to share a selection model we need a common data model. This commit
refactors PlugTableModel to provide the holistic data source for the
entire logical spreadsheet.  Proxy models are then used to isolate parts
of this data for the independent table views that make up the
presentation of this sheet in the UI.