ImageGadget : Fix crash viewing monochrome or soloed channels
In this case, we call `Tile::texture()` on the _same_ tile more than once, before calling `loadTile()` with all the textures we've gathered. If we're unlucky, in between calls, new data can arrive in `Tile::m_channelDataToConvert`, so that the next call to `texture()` will discard the previous texture. By owning a reference to the textures we prevent them from being deleted before `loadTile()` has completed and we have drawn the tile.

This fixes a regression introduced in 1c55a780d5af0a4aff32722a1b0e1d777b63ea84. Prior to that, we would call `Texture::bind()` _before_ the texture was destroyed, avoiding the crash we're fixing here. But the `Texture` could still have been deleted between the `bind()` and the drawing of the tile. It's not clear to me how we were getting away with that, but _maybe_ it could explain a rare drawing glitch we've seen during IPR?