Arnold Renderer : Work around unhelpful Arnold "inheritance" behaviour
Procedurals are the only way to introduce hierarchy into an Arnold scene, and therefore the only way to do high-level instancing. But their behaviour with respect to attribute inheritance makes them unsuitable for this task, because attributes defined on the procedural completely override any attributes defined on its children. The rules appear to be :

- `matrix` is concatenated as you would expect.
- `visibility` takes a bitwise `&` of the procedural value and the child value.
- User parameters are treated properly, with child nodes inheriting them from `ginstance` nodes referencing the procedural.
- Everything else is backwards, with the procedural's values trashing the child values. This was presumably helpful in some standin workflows prior to operators being a thing, but it is totally unsuited to our purposes.

We work around this by emulating the inheritance ourself. This has several downsides compared to hypothetical native support in Arnold :

- We can't instance procedurals if they have different attributes.
- We can't edit attributes on procedurals after they have been created. Instead we must force a complete re-expand.

I have been careful to ensure that these downsides don't apply where the only differences are in user attributes, so that we can still instance encapsulated prototypes effectively for the Gaffer Instancer node.

Fixes #3559.