NumericPlugValueWidget : Don't clobber value with invalid entry
ce15fd6 inadvertently changed `NumericWidget` validation, such that `0` was
returned for invalid input. This resulted in plugs being overwritten with
any time editing finished with invalid text in the widget. This also
had the side-effect that a `NumericPlugValueWidget` representing multiple
plugs with different values would overwrite all plugs to `0` if editing
finished without a change as the default presentation of mixed values is
as an empty string + placeholder text.

This commit reverts user-facing behaviour to that prior to `0.58.2.0`, with
the added benefit that the text field will now be updated to contain the
plug value when editing is finished, rather than the next time the widget
is updated.