StringPlugValueWidget : Don't clobber mixed values if unedited
When a widget loses focus, we receive the `editingFinished` signal,
in the case of mixed values, we were inadvertently overwriting all
plug values with the empty string from the widget.

This commit adjusts behaviour so that interaction with a mixed value
placeholder is required before the value of the plugs will be changed.

To allow the user to easily set an 'empty' string for all plugs, a
backspace press with the mixed value placeholder visible removes the
placeholder. If the widget looses focus/completes editing at this point
the plugs will be set to an empty string.

This also restores common placeholder text (if set) once interaction
occurs (and the plugs consequently share a value).