SceneViewInspector : Ensure EditWindow widgets lose focus when hidden
The mechanism by which Qt hides an active Popup mode window seems to only
apply a visibility change. This can leave child widgets with active keyboard
focus, despite being invisible. In this case, this resulted in a second
click being required to action any uncommitted edits after the EditWindow
was 'closed' by clicking outside the popup.

Just removing the last reference (and consequently deleting the window)
caused the widgets to be deleted without committing the edit (as they were
invisible?). This change ensures that if the active focus widget was within
the now invisible window, it's focus is cleared. This will result in any
PlugValueWidget's committing their edits if required.