Editor : Don't steal focus if it is already with a child
In some situations, the Editor may receive an enter event, even though
focus is already with a child widget. An example of this is if a child
has shown a popup, which is then subsequently dismissed. Qt will re-focus
the previous widget on close. This fix prevents the parent editor
stealing focus back from the child widget in this situation.