BranchCreator : Protect set hashes from eviction
They are expensive to compute but could be evicted on a regular basis when large numbers of parent paths were visited by `setHash()`, or when any other process triggered large numbers of hashes. By storing them in the global cache we make it extremely unlikely that they will be evicted, because they are competing for membership against a far smaller number of entries.

Although we should really multithread `hashSet()` as well, as the comments state, this change on its own is actually sufficient to make a huge improvement in interactive performance. The set computations are not so slow that they are unacceptable as a one-off startup cost, but are slow enough to destroy interaction if not cached effectively.