SpreadsheetUI : Disable native Qt space bar behaviour
Qt's default handling for <space> toggles the current cells selection
state, unless it's a checkbox, where it toggles the checked state.
This is a little confusing. As we already have <return> to toggle the
check state, this ensures that <space> only ever toggles the current
index's selection state.

This may not be the right long-term behaviour, but it does help facilitate
keyboard-only spreadsheet operations.