NameValuePlugValueWidget : Don't use `Plug.DynamicFlag`
The use of this to decide whether or not to show the name plug dates back to before the existence of metadata. We want to phase out the flag entirely, and reliance on the flag is the cause of #4020, so we now use only the  "nameValuePlugPlugValueWidget:ignoreNamePlug" metadata that has existed for some time.

To maintain the previous behaviour, we now register metadata for various nodes. A couple of notes on that :

- For the Options/Attributes nodes we have three choices.
	1. Register `ignoreNamePlug = True` on the base class, and override it on the Custom subclass.
	3. Register `ignoreNamePlug = False` on the base class and override it on all subclasses except Custom.
	2. Register `ignoreNamePlug = foo` on all concrete classes, but not the base class.
  Here we have opted for 1, since it represents the least code change and provides maximum compatibility with other subclasses that may exist in the wild. This matches the decision already taken with regard to the `compoundDataPlugValueWidget:editable` metadata.
- This also registers `compoundDataPlugValueWidget:editable = False` for `LightUI.visualiserAttributes` and `CameraUI.visualiserAttributes` to avoid the possibility of adding custom attributes where `ignoreNamePlug = True` may not be appropriate. This is consistent with the fact that we already provide all the relevant visualisation attributes.

Fixes #4020.