SpreadsheetUI : Conform when ValueAdators can run for compound plugs
Previously, we were using the base class `get/set` methods for
recursive handling of plugs without a `get/setValue` method.
This could allow a registered adaptor to run for set, that wasn't
invoked for canSet. We opt for this approach in this initial version
as it's less confusing - it minimises the potential locations in the
plug hierarchy where value hacking can occur.