SpreadsheetUI : Deselect row name cells after disabling rows
As disabled cell selections aren't retrievable through the selection
model, and we don't present a right-click menu, we clear any remaining
selection in the row-name column to avoid as much confusion (as possible).

The current multi-edit behaviour for row names means that we don't end up
with an awkward situation if they have selected both name + enabled cells.