SpreadsheetUI : Support drop of plugs on to cells to set input
There was some debate as to how drag/drop should interact with the sheet's
selection. This option simply replaces the selection as appropriate target
cells are hovered. This mimics the existing `setHighlighted` behaviours of
other UI elements. An alternative suggestion was to only apply the drop
to the current selection. This would simplify setting the input to multiple
cells at once, but make the case of setting a single cell harder as it
would need to be selected first. We can always revisit this over time as
it is used more in practice.