PlugAlgo : Don't omit non-persistent metadata during plug promotion
I think the original expectation was that non-persistent metadata was only for transient use by the UI, but a more important use case is the metadata created by earlier promotion, as demonstrated in `PlugAlgoTest.testPromoteInNodeConstructor()`. Just as non-dynamic plugs can be promoted, so should non-serialisable metadata be. Non-serialisability is not a sign of transience; it can just as well be a sign that restoration does not require serialisation.