Serialisation : Derive from `noncopyable`
We never intended to copy them, but had been inadvertently passing copies into the Serialiser python bindings. An upcoming commit requires python Serialisers to have mutable access to the Serialisation, so we must pass it by reference.

Note that this does mean that if a python Serialiser holds onto the Serialisation and attempts to use it later it will cause a crash. I think this is reasonable because Serialisers are pretty specialised, they are singletons, and storing a Serialisation is far outside of their design scope. The alternative would be to use smart pointers to hold Serialisation, but that implies we intend them to be shared when we don't, and impinges on natural use in C++.