Metadata clients : Use per-node metadata signals where possible
This improves performance by avoiding the many-to-many signalling we were using before. Standalone load time is reduced by around 12% for a test file with ~200 references and ~700 metadata registrations. Load time for the same file but with a NodeEditor open is reduced by 97%.

We still have some uses of the old signals, for the purposes of tracking changes to read-onliness, either for inheritance-from-parent-nodes or because we don't track things well enough to know exactly which plugs matter at the point we connect to the signals. This will need to be dealt with at some point, but I think we've got the majority of the performance improvements that are available already, and #4032 may change or approach to read-onliness anyway.