RendererAlgo : Fix context management for light filter linking
When doing parallel processing via TBB, it is our responsibility to transfer the current thread state (which includes the context) to the tasks being performed by other threads. Failure to do so could lead to set expressions being evaluated in the wrong context, and light filters not being linked appropriately.

This fixes a bug first reported on the gaffer-dev mailing list :

https://groups.google.com/g/gaffer-dev/c/wgOVxlDAkHw/m/ys3vMfXWBgAJ