SetUI : Add menu path transformation function
Production scenes often have hundreds of sets, presenting these in a flat
list is unhelpful for the user. There are however no well established
inter-facility standards for name tokenisation. As such, we allow a site to
register a transformation function that can turn a set name into a menu
path. This allows the insertion of submenus, etc. as suits their naming
convention.