Reference : Fix handling of added/removed spreadsheet columns
Rather than transfer the old children over to the new plug, we now create new children for the new plug and transfer over their values and connections using the existing transfer code. This allows us to use `addRow()` for Spreadsheets, which in turn ensures that the columns for the new child reflect any changes to the columns that have been made in the referenced file.