FileMenu : Increase number of recent files to 25
Also reserve a parameter to `addRecentFile()` so that we can improve the presentation in future.

Note that `addRecentFile()` does have a theoretical quadratic scaling issue because it uses linear search in `__recentFiles.remove()` and also saves out the config with every call. I have verified that although the new limit does increase runtime slightly, it is still only in the order of a few milliseconds.