StringPlug : Make hash more exact when we already have the value
When a StringPlug is going to perform substitutions in `getValue()` we already account for that in `StringPlug::hash()` and compute the most exact hash available. But in the case that the string doesn't have substitutions, we have been falling back to the default `ValuePlug::hash()` path which is less exact, even though we're already in possession of the exact string, and it's presumably hot in the cache from our call to `hasSubstitutions()`.

The main motivation here is actually the behaviour tested by `SceneReaderTest.testGlobalHashesUseFileNameValue()`, which could be implemented simply by calling `h.append( fileNamePlug()->getValue() )` in the SceneReader. But this broader change seemed like it was worth making because it probably has other potential upsides.