ScriptWindow : Fix broken shortcuts in floating editors
Again! This mechanism is proving to be very fragile. A better approach may be to deal with this automatically at a lower level, by having Windows know about MenuBars. Then `Window.addChildWindow()` could automatically ensure shortcuts are propagated appropriately. That's a bigger upheaval though, so for now, we use a simple workaround.