EventLoop : Fix potential issues with _UIThreadExecutor
We weren't calling the base class `QObject.event()` for events we weren't interested in handling. This prevents the base class from processing events related to the `moveToThread()` call we make. Moving to `customEvent()` means we know we're being called only for our own event type, and the QObject's event handling remains intact.