ExtensionAlgo : Fix copy/paste of exported nodes
The plugs are being exported with the `Dynamic` flag on, because that's what they need for regular serialisation. So we must turn it off again to prevent the plugs doubling up when the extension node is copy/pasted. Turning them off in the constructor is a pretty nasty workaround compared to exporting without them in the first place, but it does at least mirror what we do in `Reference::loadReference()`. Eventually we should get rid of the Dynamic flag entirely, but in the meantime it's worth making ExtensionAlgo useable, if a little ugly.

Fixes #3886