OSLImage : Avoid unecessary computes/hashes
When we don't have closure plugs, we don't need to evaluate the shading to figure out the channel names.  This commit creates an __affectedChannels plug, which is used to drive the output channel names, but also to pass through channel data when that channel is not affected by the shader.  The __affectedChannels plug can usually be computed without computing shading ( as long as closure plugs aren't involved ).  This commit also removes the channelName from the context when querying the __shading plug, avoiding duplicate hash cache entries.