Plug : Support child reordering
When children are reordered, this is automatically reflected in all output plugs. This allows reordering of the children of a promoted plug to be reflected on the internal plugs. Reordering also triggers dirty propagation, so that nodes that depend on plug order will be recomputed, as demonstrated by the new tests for CustomAttributes and Spreadsheet.