SpreadsheetUI : Clarify naming
The meaning of "section" was too overloaded. It could refer to our public sections which are grouping of columns managed by `_SectionChooser`, or what Qt sometimes calls sections - the rows/columns of the header views. Here we choose to use "column" to refer to a section of the horizontal header view, to avoid confusion when we start dealing with the sections of a vertical header view.