Replace `typedef` with `using`
This was done with a combination of `clang-tidy modernize-use-using` and manual work where Clang got confused.

Whether we use `typedef or `using` isn't particularly important in the grand scheme of things, but it is nice to be consistent, and `using` is the more modern and more flexible of the two. I'm subscribing to the idea that inside C++ is a smaller, better language trying to get out...