BackgroundTaskDialogue/FileMenu : Fix hang during save
I have no idea how I didn't catch this during testing. The last thing that
`ScriptNode::save()` does is to set the `unsavedChanges` plug to `False`.
That's a graph edit, and graph edits cancel associated background tasks,
which leads directly to deadlock if it's a background task making the edit.

The solution is to avoid associating the background task used for saving
with a `subject`, so graph edits don't cancel it. This is legitimate because
we have blocked all potential sources of edits via `BlockedUIThreadExecution`.
This also allows us to simplify the call signature of `waitForBackgroundTask()`.