Widget : Fix drag handling bug
The intention was that a widget should receive a signal `dragEnter`,
and - if accepted - a single matching `dragLeave`. But our early-out for the
candidate widget being the destination already was flawed. It failed in cases
where the widget under the mouse was a child of the destination widget,
and rejected the drag. In this case we were emitting `dragEnter` again
on the destination as we traversed up through the parent hierarchy. We now
abort this traversal as soon as we find the current destination widget.