Wrapper : Unset `PYTHONHOME` environment variable
This still prevents us getting clobbered by the external environment, but also
doesn't clobber `gdb` when it is launched by `GAFFER_DEBUG=1`. GDB links
to the system Python which can get mighty upset if used with the modules from
our internal Python.

Note : I'm not sure why we check for the existence of `$GAFFER_ROOT/bin/python`
since official builds always include it and Image Engine builds don't use the
wrapper. But I've left it for now just in case.