Prune : Fix bounds computation when filter claims a non-existent match
An alternative fix would have been to compute the local bound (by
computing the object) and merge it with the child bounds. This would
also have fixed the case where a location has an object and some children,
and then all the children are pruned away. I don't think forcing the
computation of objects in this case is prudent though, and we can fix the
more common case much more cheaply by computing childnames, which we need
to compute the child bounds anyway.

Longer term there's probably a good argument for having separate `objectBounds`
and `childBounds` plugs so that we can cheaply get the object bound alone,
without needing to compute the object.