Parent : Add `parentVariable` plug
This also has the nice side effect of halving the number of contexts created for a typical evaluation. Before, we were creating a GlobalScope in `sourcePath()` and then a second scope to evaluate the `sourcePlug` at the right scene location. This is now done with a single context allocation via the SourceScope class.