NameSwitch/Spreadsheet : Ignore unnamed rows
Previously these would be matched whenever the selector was empty, and this was proving confusing for users. Particularly so in the case of the NameSwitch, where it's all too easy to end up with an additional input that you're not using.

The new behaviour is consistent with similar behaviour such as ignoring unnamed attributes in CustomAttributes, unnamed variables in ContextVariables and unnamed tweaks in CameraTweaks etc.