Duplicate : Expose `filterPlug()`
This could not be exposed before because the BranchCreator would have been parenting the duplicates under the originals, which is not what anyone wants. But we can now default the `destination` to `${scene:path}/..` so that the duplicates appear as siblings of the original.

This also removes a quadratic scaling issue with transform computation, by computing the transforms at the same time as generating the names of the duplicates. This knocked more than 80% off the runtime of `DuplicateTest.testPerformance()`.