PlugLayout : Fix activity/summary updates
When this code was first written, we weren't using `__updateLazily()` to defer updates until the UI was visible again, so the shortcut for invisible UIs may have been warranted. But now we do defer things nicely, and we definitely don't want to lose updates for things that happened while we weren't visible.