BranchCreator : Add `destination` plug
This is evaluated in a context where `scene:path` refers to the source location matched by `filterPlug()`, allowing the user to choose a different destination for each source. This interoperates well with the new `..` syntax supported by `ScenePlug::stringToPath()` to allow relative placement, and more complex placement can be achieved using Spreadsheets or Expressions.

I debated whether we should use `scene:path` as the variable in the evaluation of `destination`, and considered other possibilities such as `sourcePath`. I chose `scene:path` for a few reasons :

- Although `sourcePath` might have been slightly clearer for nodes like Seeds, I wasn't convinced it helped at all for Parent.
- Introducing a new context variable potentially leads to more pressure on the hash cache if expressions or other inputs lead to upstream graph evaluation.
- Using the standard `scene:path` variable allows an expression to read from attributes or primitive variables at the source location, potentially providing a useful mechanism for controlling the destination. It would also interoperate well with hypothetical AttributeQuery or FilterQuery nodes that might default to querying the current location.

Multiple sources may be mapped to the same destination, in which case any clashing names are resolved using the ChildNamesMap in `mappingPlug()`. This leads to ambiguity in the `ExactMatch` result from `parentAndBranchPaths()`, which formerly filled in `parentPath`. There are now multiple potential values (one for each source path) so picking just one is meaningless so we  no longer fill `parentPath` for an `ExactMatch`. Technically this is a breaking change but I've taken the liberty because we don't need the value internally and I strongly doubt that `parentAndBranchPaths()` is used by external code, having only been made protected in `0.59.1.0` and then only for esoteric use in the Instancer. In retrospect, friendship may have been a better choice there.

The upshot of the problem above is that we can only support `processesRootObject()` when the `destination` is at the default value, so we hide `destination` on the Instancer and Unencapsulate nodes for now. We could continue to support it by requiring a further filter evaluation but I would prefer to consider alternative approaches when we have the freedom to make breaking changes.