BranchCreator : Allow `destination` to create new locations
For instance, allowing the creation of `${scene:path}/../../lights` where that location didn't exist before.

This could have been achieved by the additional of extra data structures in BranchesData, but it was already sub-optimal to have separate lookups for the destination and the associated sources. So this refactors everything into a single tree structure representing the destination locations and ancestors, holding all the information in one structure requiring only one lookup.