Context : Move frame storage to EditableScope
When using the pointer-based version of `set()`, we can't reliably determine whether or not to invalidate the hash and emit `changedSignal()` because the new value may be reusing the storage for the old value. This led to several test failures, including `SwitchTest.testIndexExpression()` and the new tests added by this commit.

We assume that `changedSignal()` will only ever be accessed for Contexts and _not_ for EditableScopes, so by ensuring that all `Context::set*()` calls make a new allocation, we can guarantee that we emit `changedSignal()` accurately. In future we intend to remove `m_context` from EditableScope and replace it with in-place storage, at which point there won't even be a `ChangedSignal` associated with an EditableScope.