TabbedContainer : Add scroll menu
This appears when there is not enough room to show all the tabs, allowing any tab to be selected using a simple dropdown menu. I've been resisting adding this for a while, because I feel fairly strongly that it only becomes necessary when the number of tabs is excessive. In the native UIs, we try pretty hard to limit the number of tabs, and the depth of nesting, with our biggest offender being the Instancer with six tabs. But even that fits reasonably well in a standard layout, so I suspect this feature request was driven more by the needs of custom nodes.

Although the standard Qt way of doing this is to use the built-in scroll buttons, I haven't used that for a few reasons :

- It doesn't give you any overview of all the tabs.
- It can take quite a few clicks to get where you're going, and you may not even know which way to go.
- The jumpy scrolling is a bit disorientating.
- The scroll buttons are placed over the tabs, making it hard to style
  them in a way which looks OK against both our raised-current-tab and
  our blending-in-to-the-background-non-current-tabs.

The implementation for this is a bit ugly, because we have to let Qt think it's using the scroll buttons for it to show the current tab properly, but it seems to work reasonably well in practice. There is a weird jump where the last tab disappears when the tab bar is so small that the height exceeds the width, but that doesn't feel too bad because there's not enough space to read the tab anyway. This appears to be down to a bug in the Qt style implementation, where it assumes that if the width is less than the height, it's dealing with a vertical tab bar.