SceneAlgo : Use `parallel_for()` in `parallelProcessLocations()`
This is an improvement on our previous strategy of managing tasks ourselves when there are many children of one location, such as with the Instancer. Instead of spawning all the child tasks from one thread, the creation of child tasks is now spread across multiple threads by the `parallel_for`. This gives a greater than 50% reduction in runtime for `FilterResultsTest.testHashPerformance()` and around a 10% reduction in runtime for `FilterResultsTest.testComputePerformance()`.

Avoiding `task` also gets us into better shape for future versions of TBB, where `task` will be deprecated and eventually removed.