Arnold Renderer : Fix bug in tracking of nodes created by procedurals
When expanding a procedural, we need to keep track of all the Arnold nodes we create so that we can report them via `procNumNodes()` and `procGetNode()`. We were doing this correctly when taking the automatic-instancing code path that we typically use, but were completely failing to account for nodes created in the non-instancing code path.