Set : Add `setVariable` plug
This passes the name of the set being processed to the `filter` input, allowing the filter to be modified depending on the name of the set. Using the existing `scene:setName` variable doesn't work for this, because even if we preserve it when pulling on `pathMatcherPlug()`, the FilterResults node will delete it internally.