GafferVDB : Use ObjectProcessor/Deformer base classes
This edges us closer to being able to deprecate and remove the SceneElementProcessor base class.

Note that this has broken backwards compatibility for the PointsGridToPoints node, because it used to apply to all locations if a filter wasn't connected. This is a historical mistake that we're trying to move away from, and I suspect PointsGridToPoints sees little enough use that the change won't be too painful.