LevelSetToMesh : Fix bounds computation
The node is useable with non-level set fields, in which case we have no idea where the isosurface might be, and a simple offset doesn't work. Even in the case of a level set, we don't know what the exterior bandwidth is so the offset is inaccurate in that case too. This is highlighted by the fact that the existing test case was had insufficient interior bandwidth for the chosen iso value, and was generating an empty mesh and non-empty bounds.